+++ src/win.h
@@ -16,6 +16,8 @@ extern char * cmd;
 extern bool icon_is_from_shortcut;
 extern wstring shortcut;
 
+extern bool cygver_ge(uint major, uint minor);
+
 extern void exit_mintty(void);
 extern void report_pos(void);
 extern void win_reconfig(void);